ortho: use a 'size_t' for 'maze.ngcells'
This pays down some technical debt from
109b794503bc48e20951a57018f0844d8508a78c, at the cost of pushing the conversion
back to `int` deeper down. This could be further propagated (through `partition`
and its callees) but we choose not to do this here in order to keep a small
manageable commit.

+++ lib/ortho/ortho.c
@@ -1530,7 +1530,7 @@ static DEBUG_FN void emitGraph(FILE *fp, maze *mp, size_t n_edges,
     fprintf (fp, "%d %d translate\n", TRANS, TRANS);
 
     fputs ("0 0 1 setrgbcolor\n", fp);
-    for (int i = 0; i < mp->ngcells; i++) {
+    for (size_t i = 0; i < mp->ngcells; i++) {
       const boxf bb = mp->gcells[i].bb;
       fprintf (fp, "%f %f %f %f node\n", bb.LL.x, bb.LL.y, bb.UR.x, bb.UR.y);
     }
